% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BEA_Regions.R
\name{BEA}
\alias{BEA}
\title{This function will take either a vector of U.S. state/territories names or abbreviations and generate a new column identifying which of the 8 Bureau of Economic Analysis Regions the statesbelongs to.
The function will automatically make state names title case and make abbreviations uppercase.
Note: OMB's division of the United States is the standard for many other Federal agencies such as EPA and FEMA.}
\usage{
BEA(x)
}
\description{
This function will take either a vector of U.S. state/territories names or abbreviations and generate a new column identifying which of the 8 Bureau of Economic Analysis Regions the statesbelongs to.
The function will automatically make state names title case and make abbreviations uppercase.
Note: OMB's division of the United States is the standard for many other Federal agencies such as EPA and FEMA.
}
