% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Census_Divisions.R
\name{census_division}
\alias{census_division}
\title{This function will take either a vector of U.S. state/territories names or abbreviations and generate a new column identifying which of the 10 Census Regions the state/territory belongs to.
The function will automatically make state/territory names title case and make abbreviations uppercase.
US territories are not part of any Census region/division.}
\usage{
census_division(x)
}
\description{
This function will take either a vector of U.S. state/territories names or abbreviations and generate a new column identifying which of the 10 Census Regions the state/territory belongs to.
The function will automatically make state/territory names title case and make abbreviations uppercase.
US territories are not part of any Census region/division.
}
