% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Census_Regions.R
\name{census_region}
\alias{census_region}
\title{This function will take either a vector of U.S. state/territories names or abbreviations and generate a new column identifying which of the 10 Census Regions the state/territory belongs to.
The function will automatically make state/territory names title case and make abbreviations uppercase.
It attempts to capture all territories, including the various Pacific Islands, by identifying a range of ways they might be named.}
\usage{
census_region(x)
}
\description{
This function will take either a vector of U.S. state/territories names or abbreviations and generate a new column identifying which of the 10 Census Regions the state/territory belongs to.
The function will automatically make state/territory names title case and make abbreviations uppercase.
It attempts to capture all territories, including the various Pacific Islands, by identifying a range of ways they might be named.
}
