% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/national_parks.R
\name{national_park_regions}
\alias{national_park_regions}
\title{This function will take either a vector of U.S. state/territories/Court District names or abbreviations and generate a new column identifying which of the 13 Federal Court Circuits they belong to (please note that the Federal Circuit is not included).
The function will automatically make full names title case and make abbreviations uppercase.
It attempts to capture all territories, including the various Pacific Islands, by identifying a range of ways they might be named.}
\usage{
national_park_regions(x)
}
\description{
This function will take either a vector of U.S. state/territories/Court District names or abbreviations and generate a new column identifying which of the 13 Federal Court Circuits they belong to (please note that the Federal Circuit is not included).
The function will automatically make full names title case and make abbreviations uppercase.
It attempts to capture all territories, including the various Pacific Islands, by identifying a range of ways they might be named.
}
