% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OMB_Regions.R
\name{OMB}
\alias{OMB}
\title{This function will take either a vector of U.S. state/territories names or abbreviations and generate a new column identifying which of the 10 OMB Regions the state/territory belongs to.
The function will automatically make state/territory names title case and make abbreviations uppercase.
It attempts to capture all territories, including the various Pacific Islands, by identifying a range of ways they might be named.
Note: OMB's division of the United States is the standard for many other Federal agencies such as EPA and FEMA.}
\usage{
OMB(x)
}
\description{
This function will take either a vector of U.S. state/territories names or abbreviations and generate a new column identifying which of the 10 OMB Regions the state/territory belongs to.
The function will automatically make state/territory names title case and make abbreviations uppercase.
It attempts to capture all territories, including the various Pacific Islands, by identifying a range of ways they might be named.
Note: OMB's division of the United States is the standard for many other Federal agencies such as EPA and FEMA.
}
